requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
Faker==9.9.1
pandas==1.3.5
asn1crypto==1.5.1
certifi==2022.6.15       
cffi==1.15.1
charset-normalizer==2.1.0
click==7.1.2
colorama==0.4.5
configparser==5.2.0      
cryptography==36.0.2     
Django
django-utils==0.0.2
Faker==9.9.1
greenlet==1.1.2
idna==3.3
joblib==1.1.0
llvmlite==0.38.1
nltk==3.7
numba==0.55.1
numpy==1.21.6
oscrypto==1.3.0
pandas==1.3.5
pycparser==2.21
pycryptodomex==3.15.0    
PyJWT==2.4.0
pymssql==2.2.5
PyMySQL==1.0.2
pyodbc==4.0.30
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
regex==2022.8.17
requests==2.28.1
scikit-learn==1.0.2
scipy==1.8.1
snowflake-connector-python==2.7.11
snowflake-connector-python[pandas]
SQLAlchemy==1.4.25
sqlparse==0.4.2
text-unidecode==1.3
textblob==0.17.1
threadpoolctl==3.1.0
tqdm==4.64.0
typing-extensions==4.3.0
urllib3==1.26.11